+++ sites/geohub/src/routes/(app)/dashboards/+page.svelte
@@ -1,8 +1,7 @@
 <script lang="ts">
-	import { browser } from '$app/environment';
 	import { page } from '$app/stores';
-	import DashboardContents from '$components/pages/home/DashboardContents.svelte';
 	import { HeroHeader, type BreadcrumbPage } from '@undp-data/svelte-undp-components';
+	import DashboardsExplorer from './DashboardsExplorer.svelte';
 
 	let breadcrumbs: BreadcrumbPage[] = [
 		{ title: 'home', url: '/' },
@@ -12,8 +11,41 @@
 
 <HeroHeader title="Explore dashboards" bind:breadcrumbs />
 
-{#if browser}
-	<div class="m-6">
-		<DashboardContents />
-	</div>
-{/if}
+<section class="solution-section">
+	<h2 class="title is-2">Discover geospatial solutions</h2>
+	<p class="description is-size-4">
+		Leverage dashboards and geospatial tools for{'\u00A0'}informed decision-making
+	</p>
+
+	<DashboardsExplorer />
+</section>
+
+<style lang="scss">
+	.solution-section {
+		margin: 64px 48px;
+
+		@media (max-width: 48em) {
+			margin: 16px 48px;
+
+			@media (max-width: 48em) {
+				margin: 48px 0;
+
+				.title {
+					margin: 0 24px;
+				}
+
+				.description {
+					margin: 0 24px;
+				}
+			}
+		}
+
+		.title {
+			line-height: 109%;
+		}
+
+		.description {
+			margin-bottom: 32px;
+		}
+	}
+</style>